requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
cycler==0.10.0
deepface==0.0.49
Django==3.1.7
dlib==19.21.1
docutils==0.16
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
gdown==3.12.2
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
ibm-cloud-sdk-core==3.6.0
ibm-watson==5.1.0
idna==2.10
importlib-metadata==3.7.2
itsdangerous==1.1.0
jeepney==0.6.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.2
mtcnn==0.1.0
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
Pillow==8.1.2
pkginfo==1.7.0
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
readme-renderer==29.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.7.2
scipy==1.5.4
SecretStorage==3.3.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
tqdm==4.59.0
twine==3.3.0
typing-extensions==3.7.4.3
urllib3==1.26.3
watson-developer-cloud==2.10.1
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.12.1
zipp==3.4.1
